pyproject.toml
[tool.poetry]
name = "wagtail_code_blog"
version = "0.0.1"
description = ""
authors = ["Dani Hodovic <dani.hodovic@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
wagtail-markdown = {git = "https://github.com/danihodovic/wagtail-markdown.git", rev = "configurable-extensions"}
wagtail = "^2.8"
pytest-gitignore = "^1.3"
wagtail-metadata = "^3.0.0"
django-stubs = "^1.4.0"
readtime = "^1.1.1"

[tool.poetry.dev-dependencies]
pytest = "^5.3.5"
pytest-django = "^3.8.0"
black = "^19.10b0"
mypy = "^0.761"
beautifulsoup4 = "^4.6.0"
django-extensions = "^2.2.8"
wagtail-foliage = "^0.1.1"
isort = "^4.3.21"
pudb = "^2019.2"
pylint = "^2.4.4"
pylint-django = "^2.0.14"
werkzeug = "^1.0.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
